+++ frontend/src/App.tsx
@@ -2,6 +2,7 @@ import {BrowserRouter, Routes, Route} from "react-router-dom";
 import HomePage from "./pages/HomePage.tsx";
 import AllTransactionsPage from "./pages/AllTransactionsPage.tsx";
 import NewTransactionPage from "./pages/NewTransactionPage.tsx";
+import DashboardPage from "./pages/DashboardPage.tsx";
 
 
 export default function App() {
@@ -11,6 +12,7 @@ export default function App() {
             <BrowserRouter>
                 <Routes>
                     <Route path="/" element={<HomePage/>} />
+                    <Route path="/dashboard" element={<DashboardPage/>} />
                     <Route path="/transactions" element={<AllTransactionsPage/>} />
                     <Route path="/newtransaction" element={<NewTransactionPage/>} />
                 </Routes>
